[MDH-30] refactor : ownerwaiting 패키지 명 수정
