+++ src/main/java/com/mdh/devtable/ownerwaiting/infra/persistence/OwnerWaitingRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.mdh.devtable.ownerwaitng.infra.persistence;
+package com.mdh.devtable.ownerwaiting.infra.persistence;
 
 import com.mdh.devtable.waiting.domain.ShopWaiting;
 import com.mdh.devtable.waiting.domain.Waiting;
